feature/add logout button

+++ src/utils/storageUtils.ts
@@ -1,12 +1,16 @@
+"use client";
+
 type StorageKey = "access_token" | "refresh_token" | "user" | string;
 
 export const storage = {
   set: (key: StorageKey, value: any): void => {
+    if (typeof window === "undefined") return;
     const data = typeof value === "string" ? value : JSON.stringify(value);
     localStorage.setItem(key, data);
   },
 
   get: <T = string>(key: StorageKey, parse = false): T | null => {
+    if (typeof window === "undefined") return null;
     const data = localStorage.getItem(key);
     if (!data) return null;
 
@@ -19,10 +23,12 @@ export const storage = {
   },
 
   remove: (key: StorageKey): void => {
+    if (typeof window === "undefined") return;
     localStorage.removeItem(key);
   },
 
   clear: (): void => {
+    if (typeof window === "undefined") return;
     localStorage.clear();
   },
 };
